[feat] 더블점프 기능 활성화 아이템 추가
바나나를 먹으면 일정 시간동안 더블점프를 할 수 있는 기능 추가

+++ Assets/Script/Player/PlayerController.cs
@@ -13,8 +13,8 @@ public class PlayerController : MonoBehaviour
     private bool isRuning = false;
     private Vector2 curMovementInput;
     public float jumpPower;
-    public int jumpCount;
-    public int curJumpCount;
+    public float jumpCount;
+    public float curJumpCount;
     public LayerMask groundLayerMask;
     public float jumpStamina;
     public float runStamina;
